Add QUERY_ALL_PACKAGES permission to manifest for Android 11 and above
Trying to get the signature of an user-installed application on Android 11 or above results in a `android.content.pm.PackageManager$NameNotFoundException` error due to [package visibility changes](https://medium.com/androiddevelopers/package-visibility-in-android-11-cc857f221cd9). Added the QUERY_ALL_PACKAGES permission to allow the app to get the signature of any installed app.